FreeTDS: Make clearly safe narrowing conversions explicit.  JIRA: CXX-3707.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@99086 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/dbapi/driver/ftds100/freetds/ctlib/cs.c
@@ -1038,7 +1038,7 @@ cs_locale(CS_CONTEXT * ctx, CS_INT action, CS_LOCALE * locale, CS_INT type, CS_V
 
 		case CS_SYB_CHARSET:
 			if (buflen == CS_NULLTERM) {
-				buflen = strlen((char *)buffer);
+                                buflen = (int)strlen((char *)buffer);
 			}
 			
 			free(locale->charset);
@@ -1051,7 +1051,7 @@ cs_locale(CS_CONTEXT * ctx, CS_INT action, CS_LOCALE * locale, CS_INT type, CS_V
 
 		case CS_SYB_LANG:
 			if (buflen == CS_NULLTERM) {
-				buflen = strlen((char *)buffer);
+                                buflen = (int)strlen((char *)buffer);
 			}
 			
 			free(locale->language);
@@ -1068,7 +1068,7 @@ cs_locale(CS_CONTEXT * ctx, CS_INT action, CS_LOCALE * locale, CS_INT type, CS_V
 			char *b = (char *)buffer;
 
 			if (buflen == CS_NULLTERM) {
-				buflen = strlen(b);
+                                buflen = (int)strlen(b);
 			}
 
 			/* find '.' character */
@@ -1119,7 +1119,8 @@ cs_locale(CS_CONTEXT * ctx, CS_INT action, CS_LOCALE * locale, CS_INT type, CS_V
 
 		switch (type) {
 		case CS_SYB_CHARSET:
-			tlen = (locale->charset ? strlen(locale->charset) : 0) + 1;
+                        tlen = (locale->charset ? (int)strlen(locale->charset)
+                                : 0) + 1;
 			if (buflen < tlen)
 			{
 				if (outlen)
@@ -1134,7 +1135,8 @@ cs_locale(CS_CONTEXT * ctx, CS_INT action, CS_LOCALE * locale, CS_INT type, CS_V
 			break;
 
 		case CS_SYB_LANG:
-			tlen = (locale->language ? strlen(locale->language) : 0) + 1;
+                        tlen = (locale->language ?
+                                (int)strlen(locale->language) : 0) + 1;
 			if (buflen < tlen)
 			{
 				if (outlen)
@@ -1152,8 +1154,10 @@ cs_locale(CS_CONTEXT * ctx, CS_INT action, CS_LOCALE * locale, CS_INT type, CS_V
 		{
 			int clen;
 
-			tlen = (locale->language ? strlen(locale->language) : 0) + 1;
-			clen = (locale->charset ? strlen(locale->charset) : 0) + 1;
+                        tlen = (locale->language ?
+                                (int)strlen(locale->language) : 0) + 1;
+                        clen = (locale->charset ?
+                                (int)strlen(locale->charset) : 0) + 1;
 			
 			if (buflen < (tlen + clen))
 			{
@@ -1167,7 +1171,7 @@ cs_locale(CS_CONTEXT * ctx, CS_INT action, CS_LOCALE * locale, CS_INT type, CS_V
 				((char *)buffer)[0] = '\0';
 			strcat((char *)buffer, ".");
 			if (locale->charset) {
-				tlen = strlen((char *)buffer);
+                                tlen = (int)strlen((char *)buffer);
 				strcpy((char *)buffer + tlen, locale->charset);
 			}
 			code = CS_SUCCEED;
@@ -1175,7 +1179,8 @@ cs_locale(CS_CONTEXT * ctx, CS_INT action, CS_LOCALE * locale, CS_INT type, CS_V
 		}
 
 		case CS_SYB_SORTORDER:
-			tlen = (locale->collate ? strlen(locale->collate) : 0) + 1;
+                        tlen = (locale->collate ? (int)strlen(locale->collate)
+                                : 0) + 1;
 			if (buflen < tlen)
 			{
 				if (outlen)
