+++ src/dbapi/driver/ftds100/freetds/tds/gssapi.c
@@ -382,7 +382,7 @@ tds_gss_continue(TDSSOCKET * tds, struct tds_gss_auth *auth, gss_buffer_desc *to
 	}
 
 	auth->tds_auth.packet = (TDS_UCHAR *) send_tok.value;
-	auth->tds_auth.packet_len = send_tok.length;
+        auth->tds_auth.packet_len = (int) send_tok.length;
 
 	return TDS_SUCCESS;
 }
